+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.7]
+### Fixed
+- split_on_adapter also defaults to both .fastq and fastq.gz from cli
+### Added
+- Options --min_qscore and --max_abs_seqlen_diff to find duplex reads
+- Default filtering on min_qscore (12) and maximum length difference in pairs_from_summary for duplex reads
+
 ## [v0.2.6]
 ### Fixed
 - Surprising behaviour in split_on_adapter to only work on gzipped fastqs by default